feat: context menu archieve and delete and  trash
