+++ client/src/components/ui/Email/ArchiveMutate/ArchiveMutate.tsx
@@ -9,33 +9,10 @@ import { PaginatedMessages } from '../TrashMutate'
 import { ToggleToolTipSpanWrapper } from '../..'
 import { useDispatch } from 'react-redux'
 import { getSelectedEmailDispatch } from '@/context'
+import { useArchiveMutate } from '@/hooks'
 
 export const ArchiveMutate = ({ disabled, threads, tip }: ArchiveMutateType) => {
-  const currentQueryKey = JSON.parse(getCookie('query:key')) || ['primary', { q: 'label:inbox category:primary' }]
-  const threadIds = threads && threads.map((item) => item.threadId)
-  const dispatch = useDispatch()
-
-  const startMutation = useMutation({
-    mutationKey: ['Archive-Message', { threadIds }],
-    mutationFn: () => modifyThread({ threadIds }),
-    onSuccess: () => {
-      queryClient.setQueryData<PaginatedMessages>(currentQueryKey, (oldData) => {
-        if (!oldData) return { pages: [], pageParams: [] }
-        return {
-          ...oldData,
-          pages: oldData.pages.map((page) => ({
-            ...page,
-            messages: page.messages.filter((message) => !threadIds.includes(message.threadId)),
-          })),
-        }
-      })
-      toast.success(`Thread has been moved to Archive!`)
-      dispatch(getSelectedEmailDispatch([]))
-    },
-    onError: () => {
-      toast.error(`Error: Thread has not been moved to Archive!`)
-    },
-  })
+  const { startMutation } = useArchiveMutate({ threads })
   return (
     <>
       <ToggleToolTipSpanWrapper
