+++ client/src/components/ui/Email/DeleteMutate/DeleteMutate.tsx
@@ -9,30 +9,11 @@ import { ToggleToolTipSpanWrapper } from '../..'
 import { DeleteMutateProps } from './DeleteMutate.types'
 import { getSelectedEmailDispatch } from '@/context'
 import { useDispatch } from 'react-redux'
+import { useDeleteMutate } from '@/hooks'
 
 export const DeleteMutate = ({ disabled, threads, tip }: DeleteMutateProps) => {
-  const currentQueryKey = JSON.parse(getCookie('query:key')) || ['primary', { q: 'label:inbox category:primary' }]
-  const threadIds = threads && threads.map((item) => item.threadId)
-  const dispatch = useDispatch()
+  const { startMutation } = useDeleteMutate({ threads })
 
-  const startMutation = useMutation({
-    mutationKey: ['delete-Message', { threadIds }],
-    mutationFn: () => trashMessage({ threadIds }),
-    onSuccess: () => {
-      queryClient.setQueryData<PaginatedMessages>(currentQueryKey, (oldData) => {
-        if (!oldData) return { pages: [], pageParams: [] }
-        return {
-          ...oldData,
-          pages: oldData.pages.map((page) => ({
-            ...page,
-            messages: page.messages.filter((message) => !threadIds.includes(message.threadId)),
-          })),
-        }
-      })
-      toast.success(`Thread has been Deleted!`)
-      dispatch(getSelectedEmailDispatch([]))
-    },
-  })
   return (
     <>
       <ToggleToolTipSpanWrapper
