+++ client/src/hooks/email/index.ts
@@ -2,3 +2,6 @@ export * from './useEmailReplyThread'
 export * from './useReplyMulti'
 export * from './useMarkAsRead'
 export * from './useToggleFavoriate'
+export * from './useTrashMutate'
+export * from './useArchiveMutate'
+export * from './useDeleteMutate'
