+++ client/src/hooks/email/useArchiveMutate/useArchiveMutate.types.ts
@@ -1,0 +1,5 @@
+import { IEmail } from 'gmail-api-parse-message-ts'
+
+export type UseArchiveMutateType = {
+  threads: IEmail[]
+}
