+++ client/src/hooks/email/useDeleteMutate/index.ts
@@ -1,0 +1,2 @@
+export * from './useDeleteMutate'
+export * from './useDeleteMutate.types'
