+++ client/src/hooks/email/useTrashMutate/index.ts
@@ -1,0 +1,2 @@
+export * from './useTrashMutate'
+export * from './useTrashMutate.types'
