+++ client/src/hooks/email/useTrashMutate/useTrashMutate.types.ts
@@ -1,0 +1,5 @@
+import { IEmail } from 'gmail-api-parse-message-ts'
+
+export type TrashMutateType = {
+  threads: IEmail[]
+}
